requirements.txt
asgiref==3.8.1
Django==5.2
django-simple-history==3.8.0
djangorestframework==3.16.0
sqlparse==0.5.3
